Version 0.6.5 Basics of QuizScreen

+++ app/src/main/java/com/napnap/testoapp/data/classes/QuizData.kt
@@ -2,7 +2,7 @@ package com.napnap.testoapp.data.classes
 
 data class QuizData(
     val name: String,
-    val completion: Double,
+    val completion: Float,
     val time: String,
     val date: String
 )
